Content models (SubClasses)
The Content model of your courses application contains a generic relation to associate different types
of content with it. You will create a different model for each type of content. All Content models will
have some fields in common and additional fields to store custom data. You are going to create an
abstract model that provides the common fields for all Content models.